+++ NINA.Plugin.SessionMetaData/NINA.Plugin.SessionMetaData/SessionMetaDataWatcher.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
 using System.Text.RegularExpressions;
+using System.Web;
 
 namespace SessionMetaData.NINAPlugin {
 
@@ -19,6 +20,7 @@ namespace SessionMetaData.NINAPlugin {
         private bool JSONEnabled;
         private string AcquisitionDetailsFileName;
         private string ImageMetaDataFileName;
+        private string AutoFocusRunsFileName;
 
         public SessionMetaDataWatcher(IImageSaveMediator imageSaveMediator) {
             SessionMetaDataEnabled = Properties.Settings.Default.SessionMetaDataEnabled;
@@ -53,13 +55,11 @@ namespace SessionMetaData.NINAPlugin {
         }
 
         private void WriteAcquisitionMetaData(ImageSavedEventArgs msg) {
-            string ImageFilePath = msg.PathToImage.ToString();
-            Logger.Trace($"ImageFilePath: {ImageFilePath}");
-            string ImageDirectory = GetImageDirectory(ImageFilePath);
-            Logger.Trace($"ImageDirectory: {ImageDirectory}");
+            string ImageDirectory = GetImageDirectory(msg.PathToImage);
+            Logger.Debug($"ImageDirectory: {ImageDirectory}");
 
             AcquisitionMetaDataRecord Record = new AcquisitionMetaDataRecord(msg);
-            string acquisitionFileNameSub = Utility.Utility.FileNameTokenSubstitution(AcquisitionDetailsFileName, msg.MetaData);
+            string acquisitionFileNameSub = Utility.Utility.FileNameTokenSubstitution(AcquisitionDetailsFileName, msg);
             Logger.Debug($"AcquisitionDetails file name: {AcquisitionDetailsFileName} -> {acquisitionFileNameSub}");
 
             if (CSVEnabled) {
@@ -99,14 +99,11 @@ namespace SessionMetaData.NINAPlugin {
         }
 
         private void WriteImageMetaData(ImageSavedEventArgs msg) {
+            string ImageDirectory = GetImageDirectory(msg.PathToImage);
+            Logger.Debug($"ImageDirectory: {ImageDirectory}");
 
-            string ImageFilePath = msg.PathToImage.ToString();
-            Logger.Trace($"ImageFilePath: {ImageFilePath}");
-            string ImageDirectory = GetImageDirectory(ImageFilePath);
-            Logger.Trace($"ImageDir: {ImageDirectory}");
-
-            ImageMetaDataRecord Record = new ImageMetaDataRecord(msg, ImageFilePath);
-            string imageMetaDataFileNameSub = Utility.Utility.FileNameTokenSubstitution(ImageMetaDataFileName, msg.MetaData);
+            ImageMetaDataRecord Record = new ImageMetaDataRecord(msg, GetImageFilePath(msg.PathToImage));
+            string imageMetaDataFileNameSub = Utility.Utility.FileNameTokenSubstitution(ImageMetaDataFileName, msg);
             Logger.Debug($"ImageMetaData file name: {ImageMetaDataFileName} -> {imageMetaDataFileNameSub}");
 
             if (CSVEnabled) {
@@ -177,6 +174,9 @@ namespace SessionMetaData.NINAPlugin {
                 case "ImageMetaDataFileName":
                     ImageMetaDataFileName = Properties.Settings.Default.ImageMetaDataFileName;
                     break;
+                case "AutoFocusRunsFileName":
+                    AutoFocusRunsFileName = Properties.Settings.Default.AutoFocusRunsFileName;
+                    break;
             }
         }
 
@@ -187,7 +187,8 @@ namespace SessionMetaData.NINAPlugin {
             public DateTime ExposureStart { get; set; }
             public double Duration { get; set; }
             public string Binning { get; set; }
-            public double CameraTemperature { get; set; }
+            public double CameraTemp { get; set; }
+            public double CameraTargetTemp { get; set; }
             public int Gain { get; set; }
             public int Offset { get; set; }
             public double ADUStDev { get; set; }
@@ -198,9 +199,15 @@ namespace SessionMetaData.NINAPlugin {
             public int DetectedStars { get; set; }
             public double HFR { get; set; }
             public double HFRStDev { get; set; }
-            public string GuidingRMS { get; set; }
-            public string GuidingRMSArcSec { get; set; }
+            public double GuidingRMS { get; set; }
+            public double GuidingRMSArcSec { get; set; }
+            public double GuidingRMSRA { get; set; }
+            public double GuidingRMSRAArcSec { get; set; }
+            public double GuidingRMSDEC { get; set; }
+            public double GuidingRMSDECArcSec { get; set; }
             public int? FocuserPosition { get; set; }
+            public double FocuserTemp { get; set; }
+            public double RotatorPosition { get; set; }
             public string PierSide { get; set; }
 
             public ImageMetaDataRecord() {
@@ -213,29 +220,42 @@ namespace SessionMetaData.NINAPlugin {
                 ExposureStart = msg.MetaData.Image.ExposureStart;
                 Duration = Utility.Utility.ReformatDouble(msg.Duration);
                 Binning = msg.MetaData.Image.Binning?.ToString();
-                CameraTemperature = Utility.Utility.ReformatDouble(msg.MetaData.Camera.Temperature);
+
+                CameraTemp = Utility.Utility.ReformatDouble(msg.MetaData.Camera.Temperature);
+                CameraTargetTemp = Utility.Utility.ReformatDouble(msg.MetaData.Camera.SetPoint);
+
                 Gain = msg.MetaData.Camera.Gain;
                 Offset = msg.MetaData.Camera.Offset;
+
                 ADUStDev = Utility.Utility.ReformatDouble(msg.Statistics.StDev);
                 ADUMean = Utility.Utility.ReformatDouble(msg.Statistics.Mean);
                 ADUMedian = Utility.Utility.ReformatDouble(msg.Statistics.Median);
                 ADUMin = msg.Statistics.Min;
                 ADUMax = msg.Statistics.Max;
+
                 DetectedStars = msg.StarDetectionAnalysis.DetectedStars;
                 HFR = Utility.Utility.ReformatDouble(msg.StarDetectionAnalysis.HFR);
                 HFRStDev = Utility.Utility.ReformatDouble(msg.StarDetectionAnalysis.HFRStDev);
-                GuidingRMS = GetGuidingRMS(msg.MetaData.Image);
-                GuidingRMSArcSec = GetGuidingRMSArcSec(msg.MetaData.Image);
+
+                GuidingRMS = GetGuidingMetric(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.Total);
+                GuidingRMSArcSec = GetGuidingMetricArcSec(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.Total);
+                GuidingRMSRA = GetGuidingMetric(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.RA);
+                GuidingRMSRAArcSec = GetGuidingMetricArcSec(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.RA);
+                GuidingRMSDEC = GetGuidingMetric(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.Dec);
+                GuidingRMSDECArcSec = GetGuidingMetricArcSec(msg.MetaData.Image, msg.MetaData.Image?.RecordedRMS?.Dec);
+
                 FocuserPosition = msg.MetaData.Focuser.Position;
+                FocuserTemp = Utility.Utility.ReformatDouble(msg.MetaData.Focuser.Temperature);
+                RotatorPosition = Utility.Utility.ReformatDouble(msg.MetaData.Rotator.Position);
                 PierSide = GetPierSide(msg.MetaData.Telescope.SideOfPier);
             }
 
-            private string GetGuidingRMS(ImageParameter image) {
-                return image.RecordedRMS != null ? image.RecordedRMS.Total.ToString() : "n/a";
+            private double GetGuidingMetric(ImageParameter image, double? metric) {
+                return (image.RecordedRMS != null && metric != null) ? Utility.Utility.ReformatDouble((double)metric) : 0.0;
             }
 
-            private string GetGuidingRMSArcSec(ImageParameter image) {
-                return image.RecordedRMS != null ? (image.RecordedRMS.Total * image.RecordedRMS.Scale).ToString() : "n/a";
+            private double GetGuidingMetricArcSec(ImageParameter image, double? metric) {
+                return (image.RecordedRMS != null && metric != null) ? Utility.Utility.ReformatDouble((double)(metric * image.RecordedRMS.Scale)) : 0.0;
             }
 
             private string GetPierSide(PierSide sideOfPier) {
@@ -302,8 +322,12 @@ namespace SessionMetaData.NINAPlugin {
             }
         }
 
-        private string GetImageDirectory(string ImageFilePath) {
-            return Path.GetDirectoryName(ImageFilePath.Substring(8)); // skip 'file:///'
+        private string GetImageDirectory(Uri imageUri) {
+            return Path.GetDirectoryName(HttpUtility.UrlDecode(imageUri.AbsolutePath));
+        }
+
+        private string GetImageFilePath(Uri imageUri) {
+            return HttpUtility.UrlDecode(imageUri.AbsolutePath);
         }
     }
 }